pyproject.toml
[project]
name = "cobrizi-api"
version = "0.1.0"
description = "API REST para Cobrizi"
readme = "README.md"
license = { file = "LICENCE" }
requires-python = ">=3.13"

dependencies = [
    "aiosmtplib==5.1.0",
    "alembic==1.18.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "bcrypt==4.0.1",
    "cffi==2.0.0",
    "charset-normalizer==3.4.5",
    "cryptography==46.0.5",
    "dnspython==2.8.0",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "fastapi==0.135.1",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "idna==3.11",
    "Mako==1.3.10",
    "MarkupSafe==3.0.3",
    "num2words==0.5.14",
    "passlib==1.7.4",
    "pillow==12.1.1",
    "psycopg2-binary==2.9.11",
    "pyasn1==0.6.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic_core==2.41.5",
    "python-dotenv==1.2.2",
    "python-jose==3.5.0",
    "python-multipart==0.0.22",
    "reportlab==4.4.10",
    "rsa==4.9.1",
    "six==1.17.0",
    "SQLAlchemy==2.0.48",
    "starlette==0.52.1",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "uvicorn==0.41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.11.0",
]

[project.urls]
Repository = "https://github.com/JulianTV123/cobrizi-api"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
